+++ bank/src/main/java/com/sophos/bank/repository/transactionRepository.java
@@ -5,11 +5,12 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @Repository
 public interface transactionRepository extends JpaRepository<transaction,Integer>{
 
     @Query("from transaction tran where tran.targetProduct.productNumber = ?1")
-    List<transaction> findAllByTargetProduct(long targetProduct);
+    List<transaction> findAllByTargetProduct(BigInteger targetProduct);
 }
