+++ bank/src/main/java/com/sophos/bank/service/productService.java
@@ -3,6 +3,7 @@ package com.sophos.bank.service;
 import com.sophos.bank.entity.client;
 import com.sophos.bank.entity.product;
 
+import java.math.BigInteger;
 import java.util.List;
 import java.util.Optional;
 
@@ -12,7 +13,7 @@ public interface productService {
     public Optional<product> getProductById(int productId);
     public boolean deleteProductById(int productId);
     public product updateProduct(product product);
-    public List<product> getProductByProductNumber(long productNumber);
+    public List<product> getProductByProductNumber(BigInteger productNumber);
     public List<product> getProductByOwnerAndGmfexempt(client client, boolean exempt);
     List<product> findAllByOwner(Optional<client> clientById);
 }
